Avanzamento nella PreferencesActivity. Cambiata la distanza da float a int. Cambiamenti minori.
# Model API
* Le distanze sono espresse come interi
* Adeguamento dei test e dei serializer

# Android
* Aggiunti RadiusSliderDialogPreference e RankSliderDialogPreference
* Aggiunte le impostazioni di default in Setup

+++ AroundMe Model API/src/it/unisannio/aroundme/model/User.java
@@ -124,7 +124,7 @@ public abstract class User implements Model {
 	 * 
 	 * @see Position#getDistance(Position)
 	 */
-	public double getDistance(User u) {
+	public int getDistance(User u) {
 		Position mine = getPosition();
 		Position hers = u.getPosition();
